+++ CryptoScanner.Data/Models/CoinModel.cs
@@ -13,7 +13,7 @@ namespace CryptoScanner.Data.Models
         public int Id { get; set; }
         public string? Name { get; set; }
         public string? Image { get; set; }
-        public int? CurrentPrice { get; set; }
+        public decimal? CurrentPrice { get; set; }
         public double? PriceChangePercentage24H { get; set; }
     }
 }
